% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/92_alignments2parsedData.R
\name{alignments2parsedData}
\alias{alignments2parsedData}
\title{What function does (short)}
\usage{
alignments2parsedData(alignment.dir, Annotation, outputDir, level,
  min_probe_number, package_seed)
}
\arguments{
\item{alignment.dir}{path to the directory that contains alignment files}

\item{Annotation}{regerence genome annotation \code{data.frame}}

\item{outputDir}{output directory}

\item{level}{character string, can take values "gene" or "transcript", 
specifying if new probe sets should be formed per gene or per transcript}

\item{min_probe_number}{minimal number or probes in a probe set}

\item{package_seed}{seed generated by function 
\code{makeOriginalPackageObject}}
}
\value{
a \code{"output type"}.
}
\description{
Short description string here
}
\details{
Long description string
}
\seealso{
\code{function or class name}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
