% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_alignments2probesetsRaw.R
\name{alignments2probesetsRaw}
\alias{alignments2probesetsRaw}
\title{What function does (short)}
\usage{
alignments2probesetsRaw(Alignments, Annotation,
  level = "Input_gene_or_transcript")
}
\arguments{
\item{Alignments}{object of class \code{Alignments}, 
prodused by \code{processAlignments} function}

\item{Annotation}{\code{data.frame} reference genome annotation}

\item{level}{\code{character} string, equals  \code{"gene"} 
or \code{"transcript"}}
}
\value{
returns \code{data.frame}.
}
\description{
Short description string here
}
\details{
Long description string
}
\seealso{
\code{function or class name}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
