% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{Alignments_class_example}
\alias{Alignments_class_example}
\title{Example of an object of class \code{Alignments}}
\format{\code{Alignments} class example object}
\usage{
Alignments_class_example
}
\value{
object of class \code{Alignments}
}
\description{
Example of an object of class \code{Alignments}, 
with slots \code{samGenome} and \code{samTranscriptome}
subsetted for the sake of space
}
\keyword{datasets}
