% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_annotate.G.alignment.R
\name{annotate.G.alignment}
\alias{annotate.G.alignment}
\title{annotates probes genomic alignment}
\usage{
annotate.G.alignment(GAlignment, Annotation)
}
\arguments{
\item{GAlignment}{\code{data.frame} of the raw genomic alignment (output of \code{processAlignments} function)}

\item{Annotation}{reference genome annotation,  \code{data.frame}}
}
\value{
a \code{"data.frame"}.
}
\description{
annotate probe genomic alignment with \code{CIGAR = 25M} only
}
\details{
Raw genomic alignment \code{data.frame} subsetted for mappings with \code{CIGAR = 25M} only,
these are mapped to the genomic coordinates of exons from the Reference Annotation \code{data.frame}.
A probes that falls in the region of annotated exon, is mapped to all transcripts corresponding to the exon.
}
\seealso{
\code{processAlignments}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
