% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_4_annotate.non25M.genomic.transcript_wise.R
\name{annotate.non25M.genomic.transcrits_wise}
\alias{annotate.non25M.genomic.transcrits_wise}
\title{Map non-25M genomic mapppings to transcripts coordinates}
\usage{
annotate.non25M.genomic.transcrits_wise(Alignment.genomic.non25M, Annotation)
}
\arguments{
\item{Alignment.genomic.non25M}{\code{data.frame} of raw genomic alignment, subsetted to non-25M mappings (\code{CIGAR != "25M"})}

\item{Annotation}{\code{data.frame}, reference genome annotation, must contatin \code{feature} column, and feature names must contain \code{"transcript"} or \code{"RNA"}}
}
\value{
returns a \code{data.frame}
}
\description{
Map genomic alignment coordinated to transcript coordinates
}
\details{
Mappings in the genomic alignment \code{data.frame} are devided into two parts:
those that have \code{CIGAR == "25M"} and those that have  \code{CIGAR != "25M"}.
The first ones are mapped to exon coordinated from the Reference Annotation,
and the second ones are suspects to be exon-exon border mappings,
which is why they are mapped to transcript coordinates.
}
\seealso{
\code{annotate.G.alignment}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
