% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_1_annotation2fasta.R
\name{annotation2fasta}
\alias{annotation2fasta}
\title{write probes in FASTA}
\usage{
annotation2fasta(originalAnnotationDir, celFilePath = "InputFilePath", author,
  email, organism, species, outputDir, reverse = FALSE)
}
\arguments{
\item{originalAnnotationDir}{directory that contains annotation files: 
\code{".pgf"}, \code{".clf"},  \code{".probeset.csv"},  
\code{".mps"} and  \code{".mps"} files, if annotation is PGF-based;
\code{".cdf"} -- Chip Definition file, \code{".CEL"} -- example CEL file, 
\code{"probe_tab"} -- tabulated .txt probes file}

\item{celFilePath}{path to CEL file, if not present in 
\code{originalAnnotationDir} and annotation is CDF-based}

\item{author}{\code{character}, name of the annotation package author}

\item{email}{\code{character} e-mail of the annotation package author 
(must contain "@")}

\item{organism}{\code{character} (ex: "Human")}

\item{species}{\code{character} (ex: "Homo Sapiens")}

\item{outputDir}{path to user-spesified output directory}

\item{reverse}{\code{logical}, if \code{reverse = TRUE}, probe sequences 
are written in reverse order (not reverse complementary)}
}
\value{
writes \code{"FASTA"} file in directory 
\code{outpurDir/Probe_Sequences_Fasta/} (created the directory if needed)
}
\description{
Fasta files from a pd.annotation object
}
\details{
Fasta files with pobe sequences generated from a parsed data object
(probe sequences and ids from original annotation).
}
\examples{
dir1 = "/Users/milchevs/ownCloud/ProbeRemapSteps14/inst/extdata/mogene2_affyLibFiles_pgfBased/"
outDir = "/Users/milchevs/ownCloud/ProbeRemapSteps14/OutPutDir/"
dir.create(outDir)
annotation2fasta(originalAnnotationDir = dir1, 
                organism = "Mouse", species = "Mus musculus", 
                author = "Vladislava Milchevskaya",
                email = "milchv@gmail.com",
                outputDir = outDir, 
                reverse = FALSE)
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
