% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{Annotation_example}
\alias{Annotation_example}
\title{Example of an annotation \code{data.frame}}
\format{Annotation}
\usage{
Annotation_example
}
\value{
example Annotation \code{data.frame}
}
\description{
Example of a reference genome annotation \code{data.frame}
}
\keyword{datasets}
