% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/91_new.R
\name{build_new_ParsedDataOld_object_PGF}
\alias{build_new_ParsedDataOld_object_PGF}
\title{build new ParsedDataOld object PGF or CDF}
\usage{
build_new_ParsedDataOld_object_PGF(level, ParsedDataOld, PS_controls,
  Annotation, ProbeSetData, outputDir)
}
\arguments{
\item{level}{"gene" or "transcript"}

\item{ParsedDataOld}{old arced data}

\item{PS_controls}{ids of control probe sets}

\item{Annotation}{reference genome annotation data.frame}

\item{ProbeSetData}{an object with new probe groupings}

\item{outputDir}{output directory, the one that was used to store Fasta files}
}
\value{
list with the same structure as original ParsedData object, containing NewParsedData
}
\description{
build_new_ParsedDataOld_object_PGF, build_new_ParsedDataOld_object_CDF
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
