% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_3_cigar2_left.R
\name{cigar2_left}
\alias{cigar2_left}
\title{cigar - to - length of the left exon mapping}
\usage{
cigar2_left(cigar_string)
}
\arguments{
\item{cigar_string}{CIGAR character string}
}
\value{
returns \code{integer}.
}
\description{
Length of the first M-region
}
\details{
If a probe maps exon-exon border region of a transcript,
it does not appear in the Genomic Alignment as a mapping with
CIGAR = "25M", but as mapping with xxM in the beginning of
the CIGAR string, and xxM in the end,
which represents length of the regions mapped 
to the left and right exons correspondingly.
The \code{cigar2_left} function gets length of the left exon mapping,
if CIGAR string has <integer>M in the left-most position.
}
\seealso{
\code{cigar2_right}, \code{cigar2length}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
