% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_cigar2_right.R
\name{cigar2_right}
\alias{cigar2_right}
\title{cigar - to - length of the right exon mapping}
\usage{
cigar2_right(cigar_string)
}
\arguments{
\item{cigar_string}{CIGAR character string}
}
\value{
returns \code{integer}.
}
\description{
Length of the last M-region
}
\details{
If a probe maps exon-exon border region of a transcript,
it does not appear in the Genomic Alignment as a mapping with
CIGAR = "25M", but as mapping with xxM in the beginning of
the CIGAR string, and xxM in the end,
which represents length of the regions mapped to the left and right exons correspondingly.
The \code{cigar2_right} function gets length of the right exon mapping,
if CIGAR string has M of the right-most position.
}
\seealso{
\code{cigar2_left}, \code{cigar2length}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
