% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_1_cigar2length.R
\name{cigar2length}
\alias{cigar2length}
\title{CIGAR-to-length}
\usage{
cigar2length(cigar_string)
}
\arguments{
\item{cigar_string}{character, CIGAR string from the alignment \code{data.frame}}
}
\value{
returns \code{integer}.
}
\description{
Converts CIGAR string of the mappint to length of the mapping
}
\details{
Converts CIGAR string to length of the mapping fragment; copied from Java script to some alignment processing tool..
}
\examples{
cigar_character = "20M1S2N5M"
mapping_length = cigar2length(cigar_character)
mapping_length
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
