% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_cleanProbeSets.R
\name{cleanProbeSets}
\alias{cleanProbeSets}
\title{filter out small probe sets}
\usage{
cleanProbeSets(ProbeSets, level, min_probe_number = 3)
}
\arguments{
\item{ProbeSets}{\code{data.frame}}

\item{level}{\code{character}, must equal "gene" or "transcript"}

\item{min_probe_number}{\code{numeric}, probe sets with >= probes will remain}
}
\value{
a \code{"output type"}.
}
\description{
Filter out probe sets that have too few probes.
}
\details{
Filter out probe sets that have too few probes.
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
