% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_getExonExonBorder.R
\name{getExonExonBorder}
\alias{getExonExonBorder}
\title{What function does (short)}
\usage{
getExonExonBorder(Als, Annotation, level)
}
\arguments{
\item{Als}{object of the class \code{Alignment}}

\item{Annotation}{\code{data.frame}, reference genome annotation}

\item{level}{character string, equals \code{"gene"} or \code{"transcript"}}
}
\value{
a \code{"output type"}.
}
\description{
finds exon-exon border probes and corresponding transcripts
}
\details{
Identifies which probes fall into a transcript region in the reference genome annotation,
and map simultaneously to the left end of one exon and to the left end of another exon of the same transcript,
based on their CIGAR string and coordinates in the raw genomic alignment.
Only mapping that have \code{CIGAR != "25M"} are considered.
}
\seealso{
\code{function or class name}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
