% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_make_objects_for_reannotation.R
\name{make_objects_for_reannotation}
\alias{make_objects_for_reannotation}
\title{Converts ProbeSets to package building input}
\usage{
make_objects_for_reannotation(Probesets, level = "gene", Annotation, shift)
}
\arguments{
\item{Probesets}{\code{data.frame}, must contain columns 
\code{"transcript_id"}, \code{"gene_id"} and \code{"probe_id"}}

\item{level}{\code{character}, must be "gene" or "transcript"}

\item{Annotation}{reference genome annotation \code{data.frame}}

\item{shift}{numeric, 
a constant that is being added to all main probe set ids, 
in order to make them larger than control probe set ids 
(and thus avoid same probe set ids when unwanted)}
}
\value{
a \code{list} with two \code{data.frame}s
}
\description{
convert ProbeSets to a format needed for package building
}
\details{
Converts ProbeSets \code{data.frame} into a list of two \code{data.frame}s:
\code{ProbeSetDataAnnotation} and \code{ProbeSetDataFrame}.
The \code{ProbeSetDataAnnotation} contains 4 colunms:
\code{fsetname}, 
\code{fsetid}, 
\code{strand}, 
\code{chrom}, where \code{fsetname} is an Ensembl gene/transcript ID.
The \code{ProbeSetDataFrame} contains two colunms: 
\code{fsetid} and \code{fid},
where \code{fid} is a probe ID from the original array annotation.
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
