% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_0_makeOriginalPackageObject.R
\name{makeOriginalPackageObject}
\alias{makeOriginalPackageObject}
\title{make original annotation object}
\usage{
makeOriginalPackageObject(originalAnnotationDir,
  celFilePath = "InputFilePath", author = "Unknown", email = "un@known",
  pkgNameSUFFIX = ".originalPackage", organism, species, outputDir)
}
\arguments{
\item{originalAnnotationDir}{directory that contains annotation files: 
\code{".pgf"}, \code{".clf"},  \code{".probeset.csv"},  \code{".mps"} and  
\code{".transcript.csv"} files, if annotation is PGF-based;
\code{".cdf"} -- Chip Definition file, \code{".CEL"} -- example CEL file, 
\code{"probe_tab"} -- tabulated .txt probes file}

\item{celFilePath}{path to CEL file, if not present in 
\code{originalAnnotationDir} and annotation is CDF-based}

\item{author}{\code{character}, author}

\item{email}{\code{character}, email}

\item{pkgNameSUFFIX}{\code{character}, suffix to package name}

\item{organism}{\code{character} (ex: "Human")}

\item{species}{\code{character} (ex: "Homo Sapiens")}

\item{outputDir}{path to user-spesified output directory}
}
\value{
a \code{"output type"}.
}
\description{
read in original array annotation
}
\details{
read in original array annotation from Affymetrix (CDF-based or PGF-based),
and produce an object required for an annotation package,
compatible with \code{oligo} or \code{affy}.
}
\examples{
alignmentDir = "/Users/milchevs/Documents/Biology/PROJECTS/Paul_Bertone/ReANNot/MoGene2/"
load("/Users/milchevs/Documents/Biology/PROJECTS/Paul_Bertone/ReANNot/Annotations/AnnotationDataFrame.Mus_musculus.GRCm38.82.RData")
dir1 = "/Users/milchevs/ownCloud/ProbeRemapSteps14/inst/extdata/mogene2_affyLibFiles_pgfBased/"
outDir = "/Users/milchevs/ownCloud/ProbeRemapSteps14/OutPutDir/"
o1 = makeOriginalPackageObject(originalAnnotationDir = dir1, organism = "Mouse", species = "Mus musculus", outputDir = outDir )
}
\seealso{
\code{function or class name}
}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
