% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_processAlignments.R
\name{processAlignments}
\alias{processAlignments}
\title{Read probe alignments from a directory}
\usage{
processAlignments(alignment.dir = NA)
}
\arguments{
\item{alignment.dir}{the path to a directory that contains genomic 
and transcriptomic alignments}
}
\value{
a \code{"Alignments"}.
}
\description{
Reads probe alignment files and creates an \code{Alignments} object
}
\details{
This funtion reads genomic alignment file (\code{Aligned.out.sam} postfix 
required)
and transcriptomic alignment file (\code{Aligned.toTranscriptome.out.bam} 
postfix requires)
and generated an object of class \code{"Alignments"}
}
\examples{
sam_file <- 
   dir(system.file("extdata",package="pdProbeRemap"), 
   pattern="example_drosophila.Aligned.out.sam",full.names=TRUE)
alignment_dir <- 
   strsplit(sam_file,  split = "example_drosophila.Aligned.out.sam")[[1]]
Als <- processAlignments(alignment_dir)

}
\author{
Vladislava Milchevskaya \email{milchv@gmail.com}
}
