% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_SAM_read.R
\name{read.sam}
\alias{read.sam}
\title{Read SAM files.}
\usage{
read.sam(myfile)
}
\arguments{
\item{myfile}{the name of the fasta file which the data are to be read from.}
}
\value{
a \code{data.frame}.
}
\description{
Reads SAM file creates a \code{data.frame}
}
\details{
This funtion reads SAM file reverts header to columns and outputs
a \code{"data.frame"}.
}
\examples{
sam_file <- 
 dir(system.file("extdata",package="pdProbeRemap"), 
                  pattern="example_drosophila.Aligned.out.sam",
                  full.names=TRUE)
SAM_list <- read.sam(sam_file)
SAM_list$x
SAM_list$header
}
\author{
Grischa Toedt, Vladislava Milchevskaya \email{milchv@gmail.com}
}
