% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{seed_example}
\alias{seed_example}
\title{Example of an object used for annotation package building}
\format{inner annotation object}
\usage{
seed_example
}
\value{
seed with example metadata information about the package
}
\description{
Example of an object used for annotation package building, 
a lisf with slot
\code{object}, that contains information about tha package
\code{parsedDataOLD}, that containns annotation, and
\code{AnnotationType}, a character string that specifies 
the formant of the main original annotation file
}
\keyword{datasets}
